+++ libraries/base/GHC/TypeLits.hs
@@ -35,7 +35,7 @@ module GHC.TypeLits
 
     -- * Functions on type literals
   , type (<=), type (<=?), type (+), type (*), type (^), type (-)
-  , type (<>)
+  , AppendSymbol
   , CmpNat, CmpSymbol
 
   -- * User-defined type errors
@@ -170,7 +170,6 @@ infix  4 <=?, <=
 infixl 6 +, -
 infixl 7 *
 infixr 8 ^
-infixr 6 <>
 
 -- | Comparison of type-level naturals, as a constraint.
 type x <= y = (x <=? y) ~ 'True
@@ -208,7 +207,7 @@ type family (m :: Nat) - (n :: Nat) :: Nat
 -- | Concatenation of type-level symbols.
 --
 -- @since 4.10.0.0
-type family (m ::Symbol) <> (n :: Symbol) :: Symbol
+type family AppendSymbol (m ::Symbol) (n :: Symbol) :: Symbol
 
 -- | A description of a custom type error.
 data {-kind-} ErrorMessage = Text Symbol