+++ testsuite/tests/typecheck/should_run/TcTypeNatSimpleRun.hs
@@ -9,8 +9,14 @@ import Data.Proxy
 tsub :: Proxy (x + y) -> Proxy y -> Proxy x
 tsub _ _ = Proxy
 
+tsub2 :: Proxy (x + y) -> (Proxy x, Proxy y)
+tsub2 _ = (Proxy, Proxy)
+
 tdiv :: Proxy (x * y) -> Proxy y -> Proxy x
 tdiv _ _ = Proxy
+
+tdiv2 :: Proxy (x * y) -> (Proxy x, Proxy y)
+tdiv2 _ = (Proxy, Proxy)
 
 troot :: Proxy (x ^ y) -> Proxy y -> Proxy x
 troot _ _ = Proxy
@@ -21,17 +27,26 @@ tlog _ _ = Proxy
 tleq :: ((x <=? y) ~ True) => Proxy y -> Proxy x
 tleq _ = Proxy
 
-tappend :: Proxy (x <> y) -> Proxy x -> Proxy y
+tappend :: Proxy (AppendSymbol x y) -> Proxy x -> Proxy y
 tappend _ _ = Proxy
+
+tappend2 :: Proxy (AppendSymbol x y) -> (Proxy x, Proxy y)
+tappend2 _ = (Proxy, Proxy)
 
 main :: IO ()
 main = print [ sh (tsub  (Proxy :: Proxy 5) (Proxy :: Proxy 3)) == "2"
+             , let (p, q) = tsub2 (Proxy :: Proxy 0)
+               in (sh p, sh q) == ("0", "0")
              , sh (tdiv  (Proxy :: Proxy 8) (Proxy :: Proxy 2)) == "4"
+             , let (p, q) = tdiv2 (Proxy :: Proxy 1)
+               in (sh p, sh q) == ("1", "1")
              , sh (troot (Proxy :: Proxy 9) (Proxy :: Proxy 2)) == "3"
              , sh (tlog  (Proxy :: Proxy 8) (Proxy :: Proxy 2)) == "3"
              , sh (tleq  (Proxy :: Proxy 0))                    == "0"
              , symbolVal (tappend (Proxy :: Proxy "abc") (Proxy :: Proxy "ab"))
                == "c"
+             , let (p, q) = tappend2 (Proxy :: Proxy "")
+               in (symbolVal p, symbolVal q) == ("", "")
              ]
   where
   sh x = show (natVal x)